Created new module common
