+++ pilot/util/common.py
@@ -1,0 +1,40 @@
+#!/usr/bin/env python
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Authors:
+# - Paul Nilsson, paul.nilsson@cern.ch, 2018
+
+import os
+
+from pilot.util.constants import PILOT_KILL_SIGNAL
+from pilot.util.timing import get_time_since
+
+import logging
+logger = logging.getLogger(__name__)
+
+
+def should_abort(args, limit=30):
+    """
+    Abort in case graceful_stop has been set, and less than 30 s has passed since MAXTIME was reached (if set).
+
+    :param args: pilot arguments object.
+    :param limit: optional time limit (int).
+    :return: True if graceful_stop has been set (and less than optional time limit has passed since maxtime) or False
+    """
+
+    abort = False
+    if args.graceful_stop.wait(1) or args.graceful_stop.is_set():  # 'or' added for 2.6 compatibility reasons
+        if os.environ.get('REACHED_MAXTIME', None) and limit:
+            time_since = get_time_since(0, PILOT_KILL_SIGNAL, args)
+            if time_since < limit:
+                logger.warning('received graceful stop - %d s ago, continue for now' % time_since)
+            else:
+                abort = True
+        else:
+            logger.warning('received graceful stop - abort after this iteration')
+            abort = True
+
+    return abort
